Add TT::MetaliumFabricProto target for external proto consumers
Currently the TT::Metalium target has a private linkage against the
fabric proto types and does not get exposed in the library symbol table.
For downstream clients, the proto files are installed, but then require
a separate recompile of the proto files to make use of the proto
classes.

This can cause a double-registration for some of those proto classes
which result in some runtime issues.

- Exports new fabric TT::MetaliumFabricProto target to fully make use of
  the proto classes under single compile and registration.
- No breaking changes for clients linking against existing TT::MetaliumFabricProto
  since it still remain as PRIVATE linkage.