Sanitize eth src and dst addresses (#34281)
### Ticket
https://github.com/tenstorrent/tt-metal/issues/34278

### Problem description
We don't check if Ethernet send source and destination addresses are
valid. This was an issue in the Debug-a-thon and what would happen is
there will be a hang on the eth core.

### What's changed
- Add a Watcher check so this issue can be caught.
- Add unit test.

### Checklist
All post commit
https://github.com/tenstorrent/tt-metal/actions/runs/20347547565
Blackhole Multi Card (same as main)
https://github.com/tenstorrent/tt-metal/actions/runs/20347550265
Blackhole Nightly (same as main)
https://github.com/tenstorrent/tt-metal/actions/runs/20347555680
cpp-unit-tests (same as main, TLB error)
https://github.com/tenstorrent/tt-metal/actions/runs/20244130614