Bilinear upsample sharding restriction and optimizations (#34797)
### Ticket
#31042
#18236
#25024

### Problem description

Bilinear upsample used to have a very strict sharding limitation, which
resulted in poor core grid usage. This was due to the fact that the
original implementation "hacked" the sliding window infra and the halo
op, and that "hack" was only correct due to the sharding restriction.

### What's changed

Added proper halo tensor creation for bilinear upsampling. The main
issue was the fact that, for bilinear upsampling, it is possible that,
given two output sticks to be computed A and B, even if the flattened
index idx(B) > idx(A), it is still possible that A requires data at
earlier positions in the input (because, if B and A are in different
rows, the col input index required for A could be higher than the col
input index required input for B, since the required input row could be
the same if they round down to the same value). Because of that,
minimums and maximums are needed to be found in the op_trace_metadata
tensor, since they do not necessarily correspond to start and end
values.

Also, reader kernel now creates a constexpr compile time LUT for
bilinear weights for any scale factors. This eliminates expensive mul
calls in the kernel

Sharding requirement lifted
Updated autoshard logic to choose more cores
Hang at #25024 fixed

Also removed some unnecessary moves in Segformer model, resulting in a
total FPS increase

171 FPS -> 211 FPS

### Checklist
- [x] Demo tests
https://github.com/tenstorrent/tt-metal/actions/runs/20458150586
- [x] Device perf
https://github.com/tenstorrent/tt-metal/actions/runs/20458167172/job/58785150680
- [x] All post commit
https://github.com/tenstorrent/tt-metal/actions/runs/20434455189
- [x] Blackhole post commit
https://github.com/tenstorrent/tt-metal/actions/runs/20434456727
- [x] Models perf
https://github.com/tenstorrent/tt-metal/actions/runs/20434440384/job/58712783020
- [x] L2 Nightly
https://github.com/tenstorrent/tt-metal/actions/runs/20460512925