Llama-3.1-8B decode TSU optimizations (#35142)
### Ticket
[Link to Github
Issue](https://github.com/tenstorrent/tt-metal/issues/33394)

### Problem description
Llama-3.1 8B decode perf has regressed since Aug. The expectation is 66
t/s/u, and before this PR it had dropped to 50 t/s/u.

### What's changed
Transformer block:
- When CCL's were switched to the fabric-based CCL's from legacy, params
were not swept, resulting in lower CCL perf. Better params for
num_links, num_workers_per_link, and chunks_per_sync were put in place.
The defaults that were in there are still used for all other models and
for prefill.
- When we implemented fabric-based all gather and reduce scatter, the
shapes we tested were mostly larger, where the amount of work being done
by each device was non-trivial. For the shapes in Llama-3.1 8B, often
times the shapes result in only a few tiles of responsibility for AG/RS
on galaxy. In this case, we max out the number of links used (to
maximize bandwidth), but as a result the num_workers_per_link maxes out
a 1. In this case, the overhead of fabric mux (which was put in place to
handle num_workers_per_link > 1) is non trivial. We've added a switch to
use fabric mux when num_workers_per_link > 1, and not use fabric mux
when num_workers_per_link == 1.

Sampling:
When measuring perf, we want to use argmax instead of the sampler. We've
added a flag to tt_sampling.py to force argmax when the default sampling
params are passed for Llama3.1-8B.

Result:
Llama-3.1 8B decode is at 66 t/s/u for b32.

### Checklist

- [ ] [Stable release
workflow](https://github.com/tenstorrent/tt-shield/actions/runs/20578531481)

---------

Co-authored-by: alnah005 <salnahari@tenstorrent.com>
Co-authored-by: Alberto Perez Vicente <aperezvicente@tenstorrent.com>
Co-authored-by: handrewsTT <handrews@tenstorrent.com>
Co-authored-by: Mohamed Bahnas <mbahnas@tenstorrent.com>