ci: add missing check all green to revert workflow (#1024)
### Ticket
None

### Problem description
Revert workflow blocks on waiting for check-all-green job to complete.
This is the work flow we use in on-pr case. It is set in GitHub as a
mandatory step, but I forgot to add it to this workflow. Currently,
revert PR never finishes because it awaits the completion of forgotten
job.

### What's changed
Added necessary job.

### Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update